+++ app/src/main/java/com/interiordesign3d/di/Modules.kt
@@ -5,7 +5,9 @@ import com.interiordesign3d.ui.screen.catalogue.item.CatalogueItemViewModel
 import com.interiordesign3d.ui.screen.designer.DesignerViewModel
 import com.interiordesign3d.ui.screen.home.HomeViewModel
 import com.interiordesign3d.ui.screen.project.ProjectViewModel
+import com.interiordesign3d.ui.screen.language.LanguageViewModel
 import com.interiordesign3d.ui.screen.onboard.OnboardViewModel
+import com.interiordesign3d.ui.screen.select.SelectViewModel
 import com.interiordesign3d.ui.screen.settings.SettingsViewModel
 import com.interiordesign3d.ui.screen.splash.SplashViewModel
 import org.koin.core.module.dsl.viewModelOf
@@ -19,5 +21,7 @@ val viewModelModule = module {
     viewModelOf(::SettingsViewModel)
     viewModelOf(::SplashViewModel)
     viewModelOf(::OnboardViewModel)
+    viewModelOf(::LanguageViewModel)
+    viewModelOf(::SelectViewModel)
     viewModelOf(::DesignerViewModel)
 }